+++ src/packages/icons/__test__/Icon.spec.ts
@@ -4,20 +4,35 @@ describe('icon snapshots ', () => {
 
   it('icon snapshots', () => {
     const wrapper = mount(EIcon, {
-      name: 'loading'
+      props: {
+        name: 'loading',
+        size: 40
+      }
     });
     expect(wrapper.html()).toMatchSnapshot();
     wrapper.unmount();
   });
 
   it('icon size', () => {
-    const wrapper = mount(EIcon, {
+    const wrapper = mount(EIcon, { props: {
       name: 'loading',
-      size: '40'
-    });
+      size: 40
+    } });
 
-    expect(wrapper.find('div').attributes('style')).toContain(
+    expect(wrapper.find('div').attributes('style')).include(
       'font-size: 40px;'
+    );
+    wrapper.unmount();
+  });
+
+  it('icon color', () => {
+    const wrapper = mount(EIcon, { props: {
+      name: 'loading',
+      color: '#ccc'
+    } });
+
+    expect(wrapper.find('div').attributes('style')).include(
+      'color: #ccc'
     );
     wrapper.unmount();
   });